reorganize files for better dependency management
