+++ .github/actions/determine-version/getCurrentVersion.ts
