+++ .github/actions/determine-version/getNextVersion.ts
